[Subnet Prioritization] Move AllocationStrategy Enum from common.py back to cluster_config.py
