+++ cli/src/pcluster/schemas/cluster_schema.py
@@ -24,6 +24,7 @@ from marshmallow import ValidationError, fields, post_load, pre_dump, validate, 
 from pcluster.config.cluster_config import (
     AdditionalPackages,
     Alarms,
+    AllocationStrategy,
     AmiSearchFilters,
     AwsBatchClusterConfig,
     AwsBatchComputeResource,
@@ -94,7 +95,7 @@ from pcluster.config.cluster_config import (
     SlurmSettings,
     Timeouts,
 )
-from pcluster.config.common import AllocationStrategy, BaseTag, CapacityType, DefaultUserHomeType
+from pcluster.config.common import BaseTag, CapacityType, DefaultUserHomeType
 from pcluster.config.update_policy import UpdatePolicy
 from pcluster.constants import (
     DELETION_POLICIES,
